refactor(core): architect follow-ups -- file splits, stash rekey, Z-order helper, badge scaling
Split KiwiCore.swift and KiwiCore+ZOrder.swift under the 350-line ceiling, extract TilingEngine.rekeyStash(oldID:newID:) and performZOrderSequence, and update state/count badge scaling to scale with thick bars.
